List force-inlining explicitly for each recognized @IntrinsicCandidate
Currently no recognized @IntrinsicCandidate will be force-inlined, at
least not due simply to being an @IntrinsicCandidate.

However, unrecognized @IntrinsicCandidate methods are currently force-
inlined. If one becomes recognized later on, we would silently cease its
forced inlining. With this change, in that situation we'll instead fail
an assertion and require an explicit decision.